worker/requirements.txt
appdirs==1.4.4
aiofiles==23.2.1
asyncio-mqtt==0.16.1
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cryptography==42.0.8
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozendict==2.3.8
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
gspread==6.1.2
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwt==1.3.1
lxml==4.9.3
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==1.24.4
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
pika==1.3.2
proto-plus==1.24.0
protobuf==5.27.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.31
stock-portfolio-shared
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==3.0.3
yfinance==0.2.25
zipp==3.19.2
